TASK: Implement linting of JavaScript files via  - solves #28
Currently we cannot access xo directly, since it has no updated webpack-loader. Thus we need to implement the linting directly via ESLint. Since we prefer spaces over tabs, we cannot use the XO sharable configs, this is due to the nature of imports/extends of rules in ESLint.

We may need to revisit this topic again if a updated xo-loader is available.