TASK: always import "index" files explicitely, as this makes editing work in PHPStorm
I know that in the standard the "import /" notation is explicitely supported;
but PHPStorm is not supporting this right now. However, because many people
in our ecosystem use PHPStorm, and to reduce entry barriers, I suggest to change to explicit /index module imports for now (at least until PHPstorm
supports it).