TASK: Move to an ES6 class for the <CheckBox/> component to increase the performance of the component.
Before this commit, the component got re-rendered everytime something on the upper react level has changed, due to the fact that a function for the onChange handler got created everytime. This commit fixes this performance issue, by using a class, which can have static methods with respect to the context.