BUGFIX: warn about missing node in ContentCanvas
Currently if node for some editable property is missing in the state (e.g. a shared footer), the error will be thrown. This will be fixed with a different way to load node data in-band, but still it would make the system less fragile if we would just warn for missing data, and not throw an error.