TASK: preserve ordering of nodetype groups
It's important to preserve the ordering of `this._groups` as we can't
sort them again by position in JS (sorting logic is too complex).

To avoid adding special condition for filtering for Neos.Neos:Content
nodes when inserting content, let's add a default group of "general" to
`Neos.Neos:Document` and assume that nodetypes without a group are a
system nodetypes that don't need to be listed.