BUGFIX: ensure people can't get to the old UI when they are in the new UI
This fixes the case that you are in the new UI, go to e.g.
the media module, and then go back to the content module -
and find yourself in the OLD UI.

... this change means that as soon as people visited /neos! in their
session once, they have to log out and log back again to get to the
old UI (or use incognito mode).

Resolves #590