BUGFIX: add quick fix for https://github.com/neos/neos-development-collection/pull/1635
In the change neos/neos@c4f6461,
the default @process.contentElementWrapping on Neos.Neos:Content was changed from
@process.contentElementWrapping to the long form
"@process.contentElementWrapping.expression" (so it was possible to specify
a position).

However, this meant the Fusion path in Frontend for such an element was calculated wrongly;
It appended __meta/process/contentElementWrapping/expressionNeos.Neos:ContentElementWrapping
to the fusionPath in the DOM.

For the "old" Neos (ember-based) UI, everything works pretty much as expected (it's
quite hard to construct a scenario where this would trigger an actual bug); but
the new React UI gets confused with rendering the element when the Fusion path is wrong.


THIS IS THE QUICK FIX, the real fix (at the correct place) is in
https://github.com/neos/neos-development-collection/pull/1635.