BUGFIX: ensure node info is only updated non-recursively in most cases
This quick-fixes a pretty bad performance regression when inserting nodes on
top levels of the tree; leading to the problem that the full node tree
was sent to the client.

Generally, we still send way too much nodes to the client as the result
of a change operation; but that should be optimized in separate steps.