FEATURE: add "allowEmpty" in SelectBox and MultiSelectBox
If allowEmpty is FALSE, we prevent removing the "last" element from the list.

For Single Select boxes, if displaySearchBox=true, we force allowEmpty to
be TRUE, as otherwise the search box would never be reachable after an element
was selected.

To migrate existing usages of Multi and Single select boxes, nothing needs to
be changed.