FEATURE: Set background color for the canvas content
This feature introduces a new Setting `Neos.Neos.Ui.contentCanvas.backgroundColor`. It can be set to any valid string for the css property `background` and defaults to a white color. With this setting set the background of the content canvas can be set to prevent it flickering on a page load.

Resolves #863