BUGFIX: fix initialization errors if navigating to a different page too quickly
## How to reproduce

- go to a page with many content elements
- very quickly navigate to another page

Before this change, you often got quite some errors in the console,
because node initialization was not yet finished, even if the node
was not even displayed anymore. This also made our Integration Tests
brittle.

With this change, we always check on node initialization time if the
document has changed in the meantime.

With this fix, I also never encountered the "getComputedStyles on
undefined" CKEditor error anymore; I think this is happening as a
consequence of this bug.