!!! FEATURE: make Editor more configurable / extensible
While building https://github.com/sandstorm/BlockStyles, I found the internal
API for configuring the RTE lacking some features. Specifically, the following
has changed with this PR:

## create more explicit inline editor configuration format

The old format looked like this:

ui.aloha:
  […]
    h2: true
    h1: false

This was then server-side transformed to:

aloha:
  […]: h2

-> i.e. false values were removed automatically.

Now, I wanted to make it possible to add specific options to the inline
editor instance (more than the "TRUE/FALSE" from above).

Thus, I created a new format:

ui.inline.editorOptions:
  formatting:
    h2: true
    h1: false
  placeholder: "placeholder string"
  autoparagraph: true

… which is used internally throughout the codebase; and the old "aloha"
format is converted to the new format.

This enables to pass in arbitrary objects for configuring a specific
editor instance.


## make FrontendConfiguration available in manifests

We now load the FrontendConfiguration early and pass them into the manifest

## !!! change way Store is passed to manifest

before, the store was passed to the manifest as 2nd parameter. To make this
more extensible, we now pass an object as 2nd parameter, currently containing
store and frontendConfiguration