BUGFIX: ensure other people's personal workspaces are never shown in workspace selector
In the old UI code, this was done in TargetWorkspaceController.js#targetWorkspaces;
where a check of "isUserWorkspace" happened (which
in turn checked whether the workspace name started
with "user-").

We now replicate this logic in the backend.

I am not 100% sure why this is needed, as this might
only occur when the "OWNER" of the user workspace is NULL
(and I am not sure how this can happen - maybe when a user
is deleted but the corresponding workspace prevails?)