BUGFIX: rework Navigate component CSS to use flexbox
This removes magic numbers from CSS; and fixes styling bugs
when new components are rendered in the Navigate Component using
unplanned extensibility (e.g. like https://github.com/sandstorm/FulltextSearchInNodeTree).

It should not have any UI/style implications; I checked the old
and new UI side-by-side and took screenshots of both of them, and
I think they match pixel-perfect.

By not relying on ToggablePanel and the inner logic thereof for
auto-sizing and animation; and instead using flexbox, the robustness
of the CSS in the navigate component could be improved and the logic
simplified.