BUGFIX: make moving of nodes more robust
## Problem Description

when moving nodes, it sometimes happened that the node
on the original location was not removed.

## Root Cause Analysis

It turned out that the "Remove" Feedback (which is part
of the feedbacks returned on move) got transmitted
with the wrong contextNodePath (e.g. where the node was
*already moved*); thus it could not be removed in the UI.

For me in some cases it correctly worked; though I was
not able to track the issue down to specific cases. In
90% of the cases it did NOT work. I tested with a custom
4-column element; where moving in the columns 2-4 reliably
triggered the error. Interestingly, moving from Column1
mostly worked...

## Solution Fix

the "clone" only clones the node; without cloning the
corresponding NodeData object. The fix is to additionally
clone the NodeData object itself; thus we detach it
from persistence, thus it keeps the old NodePath (which is what we need).

Interesting this has not happened much more for people :-)