TASK: Improve "Changes" saga
In order to make behavior slightly more predictable and
avoid race conditions on the server the changes are now
sent in order and with only one request at a time.

We collection changes from change actions until one of the
following happens:

* we have taken 3 actions (to avoid too many unpersisted changes)
* 2 seconds have passed

If at that point another request is still running, the wait is
prolonged until that request has finished at which point the
new batch of changes will be persisted.