BUGFIX: Prevent data loss in change saga
The previous behavior left a window in which a change could be lost
(up to 1.5sec) due to waiting for the next opportunity to save.
Now pending changes are submitted to the server as soon as the last
save request finished. This should minimize the risk especially
in conjunction with disabling the publish button while chnages
are being persisted.

Fixes: 1933