BUGFIX: Prevent backend inside of client frame in error case
The old backend would show up inside the client frame in case a
rendering exception happened. The reason is the PageHandler in Neos
which just renders a (legacy) backend around the error message because
that was the only way the legacy backend could still work in case
of an error. This leads to the backend in backend problem and can be
avoided by a custom error handler for the new UI that replaces the
old one.