BUGFIX: Make make work on macOS (#2159)
* BUGFIX: Make make work on macOS

The original Makefile did not work for me, since webpack and cross-env
could not be found.

This change fixes it, calling those binaries like it is done for lerna.

* BUGFIX: Fix phony declaration

I could not find a single place that documented wildcards for phony
target declaration. And in fact "build" failed to be built, if run on a
case-insensitive filesystem, since "Build" exists.