BUGFIX: Limit out of band reloading to content-nodetypes (#2269)
If a property with `reloadIfChanged` is used on document level an error occured when the ui tried out of band rendering on the root page in a non-default dimension. The configuration is strange but can occur easyly via Mixins and Neos should behave better.

The reason for the error was that the ui checks the type of the parentNode as beeing a `ContentCollection` but did not validate wether the current node is `Content`. That lead to accessing the parentNode of the siteNode which fails for the siteDocument because '/sites'  does not exist in all dimensions.

The solution is tho check wether the node is of type `Content` before checking the type of the parentNode beeing a `ContentCollection`.