TASK: Upgrade webpack to version 4 (#2369)
* Task: Update webpack dependencies

Update webpack to version 4 and directly update related packages like webpack-merge, webpack-livereload-plugin and replace the incompatible extract-text-webpack-plugin
with mini-css-extract-plugin.

* TASK: Update webpack config to be version 4 compatible

* TASK: Update postcss dependencies

Updates postcss related packages:
 - postcss-css-variables
 - postcss-hexrgba
 - postcss-loader
 - postcss-import
 - postcss-nested

* TASK: Updates loader packages

Updates css-loader, file-loader and style-loader

* BUGFIX: Remove composes of reset

The composes in the class selectBox__contents of the selectbox
leads to the creation of a second reset file that overrules regular css rules.

It seems to be a bug in css-loader.

* Bugfix: Remove last pieces of ExtractTextPlugin

* TASK: Replace internal dependencies by regular one

The build essential package uses internal dependencies to pass thru the
mini-css-extract-plugin dependency. But this is just used in the extensibility and
so we use now just a plain require.