BUHFIX: Fix missing q helper on latest master in backend
The neos master registers q as a static function which is not compatible with later versions of neos.
Since some of the services in the ui rely on flowQuery and use a custom defaultConfiguration this
causes trouble when running on latest master.

As backwards compatible solution this change includes the eel default context from `Neos.Fusion.defaultContext`
and merge those with `Neos.Neos.Ui.configurationDefaultEelContext` before evaluating the expressions.

This also allows to remove the default Eel helpers from the configuration and only register the
additional helpers of the ui itself needs.