BUGFIX: Fix content operations after empty outOfBandRendering
After creating a new node with an empty renderedContent
payload from RenderContentOutOfBand lead to an issue with
the NodeToolBar.

So it was not possible to create, delete, copy or cut a node.
This patch prevent the clickOutside handling for the NodeToolBar. So we don`t close the dialog by accident.