FEATURE: Add EditPreviewDropdown to the Primary toolbar instead of the EditPreviewPanelToggler
With this change the editPreviewModes can be switched directly from the dropdown which saves lots of space
and is way shorter than opening a panel. Since the change uses the spot of the exiting EditPreviewPanelToggler
no additional place in the primary toolbar is used. Instead the now the title and edit/prebiew icon of the current mode
are visible directly so editors actually habe a chance to spot in which mode they currently are.

This change allowed to remove the EditPreviewPanelToggler and EditPreviewPanel.