FEATURE: publish transpiled (ES-Modules) version of react-ui-components
To make react-ui-components usable standalone, we
should publish a transpiled-to-es-modules version
of the code, in the same manner as many other projects
publish their library code.

This way, the react-ui-components are embeddable
e.g. using parcel.js, or via webpack; without setting
up typescript compilation of node_modules (which is
discouraged).

NOTE: the .postcssrc file in react-ui-components
      is necessary to indicate to other bundlers
      (like parcel) that our code needs exported class
      names when doing an
      "import style from './style.css'".