BUGFIX: Allow a contentcollection to be the insertion anchor
This is a fix for https://github.com/neos/neos-ui/pull/2609
which now allows to add the insertion-anchor attribute to
Neos.Neos:ContentCollection by default and use it at any
position in Fusion/AFX for rendering and adding nodes.

Previously the contextpath attribute of a content collection
would break this mechanism as it was checking that only the
outest parent has the contextpath.