BUGFIX: Focus first form element in node creation dialog (#2708)
* BUGFIX: Fix focus handling for text fields

The property autoFocus is never used and therefore the focus is never set for the text field.
Also fixes the reference handling!

* TASK: Make Dialog componet not by default focused

The dialog was always focued by default and therefore the focus of the
inner elements did not work.

* BUGFIX: Relsove datatype error for EditorEnvelope

The option property should be an object and an array was used.