BUGFIX: Avoid URLs to be rendered twice in link- and reference editors
This removes some ancient hack that tried to determine the "uriInLiveWorkspace".

Background:

The code extracted the context path that was previously specified via
`@<context>` in the "backend URLs". Due to the dedicated preview routes
that were introduced with Neos 5.1 this no longer worked and led to the
preview URL rendered twice instead..

Related: neos/neos-development-collection#2965