BUGFIX: Manually adjust versions (#2751)
The versions of build-essentials and positional-array-sorter are wrong and does not fit the version of the current release. That leads to the problem that in the upcomming release the version will be not replaced. The new release only replaces the new version number with the latest released version. So the two packages will never get updated.