BUGFIX: Speed up node move actions
WIth this change instead of returning the
information for the moved node and all its
descendants only the changes to the
new and old parent and the information
about the old and new context path
are submitted.

The already stored nodes are then updated
with this information instead of being replaced
completely with the new information from
the server.

This improves the response time by ~50%
and memory consumption by 25% - 80%.
Large move operations still need a lot of
memory and some time to compute
the updated workspace information.