FEATURE: Use language plurals in I18n registry  (#2827)
* FEATURE: Use language plurals in I18n registry

Adds the quantity parameter to the translate function of the I18n registry. So it is possible to render the language labels based on the quantity. If a label is not available as plural but we passthu a quantity we just get the singular of the requests label.

So this change will not break existing code.

* TASK: Use language plurals for the tabs validation