FEATURE: Improve help of the make command
The old make help missed some commands and was not very nice, as the spacing for long commands did not exist and therefore the command floats into the description.

With these change we get all commands rendered and can define the space between the commands and the description. Additionally we have a colorful output.