FEATURE: Expose parentNode for ClientEval in right side bar (#2989)
* FEATURE: Expose parentNode for ClientEval in right side bar

In some circumstances, we need to use ClientEval to hide a property, for instance. Actually it is just possible to use the current node data or some child information. But we have not much information about the parent.

As the neos-ui knows which node is the parent, it is quite helpful to use also parent node information in ClientEval. This PR adds the parent node to the context so that we have `node` and `parentNode` available in ClientEval.

* TASK: Resolve linting issues