TASK: Disable inline editing for read only workspaces
When a workspace is read only, we want to avoid being able to edit content. Therefore, we disable the inline editable mode. Additionally, it was possible to focus the nodes in the guest frame via the content tree. In this case, the node toolbar was visible, and we need to hide this as well.
We cannot remove it, as we then need to reimplement the scroll to node behavior.