FEATURE: Introduce icon configuration for nodetypes
If you use custom icons, you want to use detailed icons, which describes the use case better, in the node type model. But then the detailed icon is not the best choice for the document or content tree.

So, this feature will bring a new configuration to have a different icon for the node type dialog tiles. We can now configure a icon and a largeIcon setting for the node type.