FEATURE: Adds restore button for user impersonation
With Neos 8.0 we introduce the impersonation feature. So, it is possible to impersonate as a different user from the user management module.

This change will add related endpoints to the neos-ui to be able to communicate with the Neos.Neos API. The impersonation state will be saved in the Redux store, and a saga handles the restore of the origin user.

Besides the data handling stuff, we also add a new button to the user menu.

Related Neos PR:
https://github.com/neos/neos-development-collection/pull/3648