BUGFIX: Repair data views for the TimeSeriesView (#3112)
* BUGFIX: Resolve type error in Hero and Column view of the UI

The hero and column view expect the value property to be a string. Some packages provide number values and that leads to type errors.

* BUGFIX: Adjust compatibility with the d3scale package

We have updated d3scale and in version 3.3 of the package the handling for the ticks function has been changed. With this change the ticks function is working again.