TASK: Temporary disable of 2 dimension tests
We have some issue on CircleCI since the PHP8 image. The switch of the dimension fixture is not working and then the 2dimension test is always running on the data of the 1 dimension fixture. As we have just one test case for the 2 dimensions, and it works locally, we disable that for a while and investigate the issue.