BUGFIX: Load content tree without waiting for guest frame
This way the content tree works even
when the guest frame crashes.
The information from the guest frame is not needed anymore since #2927 to build the content tree.
This also make the UI more responsive as the content tree now loads at the same time as the document tree.

It is still necessary sometimes to update the current document node when the guest frame is initialised as the actual rendered page might target a different document node than the one expected.

Resolves: #3251