TASK: Use yarn versions to increase the version numbers (#3279)
* TASK: Make yarn version command work

To be able to use yarn version we need to configure a version in the main package.json as well and define a changesetBaseRefs as we are not using main or master as default branch.

* TASK: Use yarn versions to increase the version numbers

Since we removed lerna, we need to change the version numbers in another way. Yarn offers the version plugin.