BUGFIX: Fix dropdown spacings for reset button
Since the esbuild change, the dropdown reset button is moved into the space from the chevron. This did not happen before because the styles from dropdown and the reset were combined with composes. Right now, we don’t use composes anymore and the rewrite had issues here so that padding clashed a bit.