TASK: Handling extensibility module path
The extensibility module path need only be changed to the dist folder for the NPM packages. The regular code base need to target the src folder. So we now have the adjust-extensibility command with a parameter and we switch the module path before the NPM Release and switch back to src before the git commit.