FEATURE: Implement Api Framework
This implements the basis for the Neos UI API, by establishing
Invariants as well as JSON-Serialization/-Deserialization procedures for
the following Primitives:

A `Command` contains an instruction for performing a change through the
application. It is handled by a corresponding `CommandHandler` and
throws an Exception if anything goes wrong.

A `Query` represents a question that the application can be asked. It is
handled by a corresponding `QueryHandler`, which produces a
`QueryResult` and throws an Exception if anything goes wrong.

A `Notification` contains information about an event that happened
outside the usual Request/Response cycle of the application.
`Notification`s can be sent to one or more `Inbox`es. `Inbox`es can be
polled for recent `Notification`s.

A `Status` contains state information that can be actively pushed from
the server to the client outside the usual Request/Response cycle of the
application. `Status`es can be updated for one or more `Inbox`es.
`Inbox`es can be polled for each current `Status`.

A `DTO` (Data Transfer Object) is a plain PHP object that follows a set
of specific rules:

- It must extend the provided `Dto` class
- It must have a public constructor
- It must declare all its public properties as promoted, readonly
constructor parameters
- All its public properties must have one of the following types:
  - `string`, `int`, `float`, `bool`
  - `mixed`
  - Class that extends `Dto`
  - Class that extends `ListDto`
  - A union of any of the above
- It must be immutable

`DTO`s are used to model arbitrary, serializable data structure for
exchange between client and server.

`Command`, `Query`, `QueryResult`, `Notification` and `Status` are also
`DTO`s.

A `ListDTO` represents collections of `DTO`s. It is itself a plain PHP
object that follows a set of specific rules:

- It must extend the provided `ListDto` class
- It must have a public constructor
- It must have exactly one, variadic constructor parameter `$items`
- The constructor parameter `$items` must have one of the following
types
  - `string`, `int`, `float`, `bool`
  - `mixed`
  - Class that extends `Dto`
  - Class that extends `ListDto`
  - A union of any of the above

Every `ListDTO` serializes to a plain array.