FEATURE: Implement `@neos-project/framework-observable` package
The `@neos-project/framework-observable` package provides a very simple
implementation of the `Observable`-pattern.

Furthermore, based on `Observable`s, it establishes two key primitives
for use in the UI:

A `State` observable represents a value that changes over time. It can
be asked about its current value and any subscriber will receive the
current value immediately.

It also provides an `update` method, that allows to push state changes
calculated from the current state.

A `Channel` observable represents a sort of message bus. A subscriber
will receive all messages that have been published to the `Channel`
since the subscription was established.

It provides a `publish` method to push messages onto the channel.