BUGFIX: Colocate react-ui-components esm-build declarations and sources
The relative import in the selectbox component forced
typescript to create a different file structure for the
source files than for the declaration files causing issues with importing the esm build.

By using the package style import the folder structure is correct again and by defining the rootDir in the tsconfig file an error will be thrown if a relative include outside of the components package is used.

Resolves: #3329