BUGFIX: Adjust width of dropdown content
The content of the user dropdown was limited in the width, and therefore it could happen that the space for the labels was not enough. Now the contents section can get bigger than the trigger button.

Was also a issue in 7.3 and not an issue related to the impersonation feature.

Fixes: #3235