BUGFIX: Change context from pull_request to  pull_request_target
GitHub Actions typically run in the context of the repository where they are defined, so by default they won't have access to forks of the repository.

In order to enable your GitHub Action to work for PRs based on a fork, you can use the pull_request_target event instead of the pull_request event. The pull_request_target event is triggered whenever a pull request is opened, edited, or synchronized on a forked repository, and it provides a head_repository context that you can use to access the forked repository.