BUGFIX: Image cropper has incorrect height with 150% zoom
The `containerHeight` should be (on my pc) `532px` but instead is higher with `617px` as the `ReactCrop` component will already be rendered and falsifies the result (stretches the inner height).
To get the correct height of an empty SecondaryEditor, we either need to calculate `100vh - 41px - 41px` (the size of the upper toolbars)

An alternative approach would be, to `display: none` everything in the secondary editor (especially the `ReactCrop` component) to be able to calculate the correct height without any inner influence, and later show the inner contents again.