BUGFIX SpeedUp Inspector Title E2E Test
Previously do to some automatic magic tescafe would wait before clicking into the iframe because the unappliedChangesOverlay is already in the dom and this warning is shown:

> TestCafe cannot interact with the <iframe name="neos-content-main"> element because another element obstructs it.
  When something overlaps the action target, TestCafe performs the action with the topmost element at the original target's location.
  The following element with a greater z-order replaced the original action target: <div class="unappliedChangesOverlay">.
  Review your code to prevent this behavior.

This waiting results in a super slow down of multiple seconds!
To fix this we dont attempt to click directly in the iframe via selector but we use click left to the inspector (via negative offset) that allows to trick tescafe.

Alternative options are possibly to use a ClientFunction, absolute coordinate clicking (doesnt seem possible with testcafe) or adding a timeout `Selector('iframe', {timeout: 200})` to the selector, which would still cause the warning to dispatch though.