BUGFIX: Resolve issue with custom data source displaying additional l… (#3701)
* BUGFIX: Resolve issue with custom data source displaying additional labels incorrectly

Previously, there was an issue with the custom data source where the secondary label, utilized with label and preview, was not being displayed correctly. This commit addresses the problem, ensuring proper alignment and display of the secondary label in conjunction with the label and preview elements. The implementation includes the use of a CSS grid layout to achieve consistent and accurate rendering of the custom data source information.

Fixes: #3675

* BUGFIX: Adjust styles for long labels

Long labels float out of the container and do not look nice. This adds an ellipsis and the tooltip is always available on hover to read the whole text. The Image is now also aligned in centered position.