FEATURE: Create `<PublishDiscardDialog/>`
The PublishDiscardDialog is a modal that is supposed to pop up whenever
an editor starts publishing or discarding their workspace.

Editors need to confirm their intent, after which they are presented
with a process indicator that blocks the screen as long as the remote
operation is not finished yet.

Finally, Editors see a success message or an error message - depending
on the outcome. In both cases, they need to acknowledge the info. If
they see an error message, they are also given an opportunity to retry
the action.

This commit creates the visual implementation of the dialog but does not
wire it to the application state yet.