TASK: Create backend mechanism to collect information about rebase conflicts
Conflicts are generated from the CR-side `WorkspaceRebaseFailed` exception.

A conflict consists of a `TypeOfChange` (generated from the original CR
command that failed) and a `ReasonForConflict` (generated from the
exception that was the cause for the command to fail). It also contains
information about the affected node, the associated document and the
associated site.

Conflicts are squashed. Only the first conflict per NodeAggregateId will
be registered. This is to reduce visual redundancy, as UI users do not
need to know about every single failed constraint check. They only need
to understand what change they are going to lose, if they're choosing a
destructive action to resolve the rebase conflicts.

Also:
The SyncWorkspace command has been moved to a dedicated namespace within the
Application layer. The `rebaseWorkspace` action has been renamed to
`syncWorkspace` everywhere to improve naming consistency.