BUGFIX: Allow building UI with FA6
There are some issues with the updated TS types from FA 6 which don’t work with out code.
We ignore the issue for now, but could either find a fix in this PR or separate the problem.
With the change the UI builds fine and the new icons can be used.